Added function to migrate config file
If there is a config file in ~/.config/openghg, 'migrate_config'
will move it to ~/.ghgconfig/openghg.
If ~/.config/openghg/openghg.conf does not exist, 'migrate_config'
raised a FileNotFoundError.

If 'read_local_config' finds that there is no config at the path
returned by 'get_user_config_path', it will try to call
'migrate_config'; if this fails, it re-raises the FileNotFoundError
with a message telling the user to run '--quickstart'.

Tests added to check if 'migrate_config' successfully moves the
config file and removes the 'openghg' directory from '~/.config',
as well as checking if the assertion is raised if the config file
is not found.