Modified create_config to try migrate_config
If no configuration file is found in ~/.ghgconfig/openghg,
create_config will try to migrate_config; if this fails, it will
run the old code to create a new config.

I changed the logic slightly, since migrating the old config
didn't assign a value to the 'config' variable. This variable
is set in the other two cases; instead of using a flag to write
the config at the end, I had to write at the end of the two
"non-migrate" cases.

I added an extra test to check that create_config will actually
migrate the old config and log it.