Fixed typos and formatting for ancillary data tut.
Link to "transform" module is hardcoded; I'm not sure
if there is a way to have a formatted link to other parts
of the documentation; the only method I could find only
works for extenal links, so I linked the docs on openghg.org.

I added footnotes to remove some of the parenthetical notes,
as well as to define some of the acronyms.

A link to an article about CAMS was added; I couldn't find a CAMS
homepage to link to.