Bools now passed to metastore directly
Previously, booleans were converted to strings
before passing them to the metastore, but now
the conversion of booleans is handled by TinyDB
directly (i.e. by Python's JSON module).

Thus Boolean metadata passed to the metastore will
be returned as Boolean in searches.

To facilitate this, I had to parse the attributes
added to the xarray Dataset that is written to netCDF.
Any metadata value that is not int, float, str, or List
is converted to str. (This includes bool, which is a subtype
of int, but we make sure to convert these to strings.)
This change happens in `assign_data`.

In `datasource_lookup`, I modified the parsing of the required
metadata so that string values are converted to lower case, but
other types of values are not modified.

Some possible improvements:
1) "start_date" and "end_date" are converted to strings.
This doesn't seem to cause problems, but is maybe inconsistent
with the intent of this commit.

2) The attributes for the xarray Dataset are modified in `assign_data`.
It would be better to modify this dataset first; then `assign_data`
would only be responsible for checking if a new datasource is needed
and adding the data to the object store. (Rather than formatting the
data as well.)