Formatting, removing python 3.9 dependence
The merge operator `|` was introduced for
dicts in Python 3.9; I was using this in `search_footprints`
but have changed to using `update` instead.
To give priority to the metadata in `args`, we need to
update `kwargs` with `args`.

Other changes include removing an irrelavent comment I added
earlier, and some changes to satisfy flake8.