Removed unused import
I commented out a line with `to_lowercase` in `_base.py`
and forgot to remove the import of `to_lowercase`.

All tests pass without the `to_lowercase` step.
Formatting of metadata currently happens in many places,
we should refactor this so that metadata formatting is
controlled in a single place (if possible). Until then,
it seems best to leave this old piece of code commented out.