Created _connection module in openghg.store
This module implements most of the functionality
from BaseStore related to the metastore.

It has a subclass registry and a factory function
for creating connections for certain datatypes
given a string for the data type (e.g. "emissions").

Subclasses mainly just redefine class variables
related to the location of the metastore for that
data type, as well as required and optional
keywords for storing data of a given type.