Modified Emissions `read_file` to use connection
To avoid conflicts with the context manager in
the Emissions class, I made the `__enter__` method
close the metastore, and stopped `__exit__` from
writing to the ._data file in Emissions/uuid.

I only modified `read_file` for this commit.
It passes all the tests for `read_file `except
for one: `test_read_file_fails_ambiguous`.
(It fails `test_add_edgar_database`, but this
uses `transform_data`, which I haven't updated.)

I don't think this test should fail: the parse
function adds the correct database version to
the metadata, so the missing metadata for the
file that is read from `test_datapath3` is
added before it goes to `datasource_lookup`.

I added debugging print statements that show this.

I'm not sure why this test failed in the previous
version of `read_file`. The only part of the
old code that I omitted was adding attributes to
the Dataset in `assign_data` (now called `add_to_store`).
But this wouldn't overwrite any existing attributes,
so it wouldn't change that `database_version` is set
to `v50` in the metadata of the third test file.

Whether or not it should fail depends on whether
we want to use metadata pulled from the attributes
of the data file to