Created Metadata, MetadataContainer classes
Metadata objects have a key and value, plus
a string valued tag "origin", which is used
to track where metadata comes from, which
should help with reconciling different pieces
of metadata.

MetadataContainer objects essentially act
as dictionaries of metadata, but they introduce
the option to format metadata, and dispatch
subsets of metadata for different purposes.