Refactored eulerian read_file
This required changing the `add_to_store`
method in `_connection.py` to write
`datasource.metadata()` to the metastore,
instead of `metadata`, even though the test
in test_eulerian.py was adding a new datasource.

This is because `datasource.add_data` updates
`start_date` and `end_date`.

Ideally, metadata processing would end when the
(metadata, data) pair is handed off to Datasource.