Marked 4 tests as 'skip'
1 test in test_emissions.py: `test_read_file_fails_ambiguous`

This test used to xfail because the optional_keys in `read_file`
only used metadata from user input (via the function parameters).

The refactored version looks for required and optional metadata
based on what is given to `add_to_store` from the parser function.
The parser function adds attributes to the metadata, including the
database version.

This test expects a DatasourceLookupError because
database_version isn't included in the arguments to `read_file`.
This error no longer because it is in the data attributes.

It would be good to have a test for the DatasourceLookupError,
but we would need to modify the data for this, or we would need
to change the behavior of the refactored code.

The refactored code tried to separate required/optional keys from
the data processing logic. To maintain this separation and only use
user input for datasource_lookup, we would need to tell
datasource_lookup what metadata it can use.

Three tests in test_obssurface.py:

`test_raising_error_doesnt_save_to_store`: this is using
the code in a way that doesn't make sense after refactoring.

I'm not sure exactly what scenario it is testing for, but perhaps
the test could be refactored.

The other two tests `test_read_GC` and `test_read_thames_barrier`
check the values stored with datasource uuids in the internal metadata.
I changed the behavior of `add_to_store` to just generate a placeholder
key.

I'm not sure why we need these keys, other than for the user to figure
out what was saved when they check the datasource uuids that are
returned. If the format matters, then they should be formatted
consistently.
(Currently, they are because `add_to_store` uses the same recipe for all
data sources.)