Added compression to Datasource.save()
I also added a test to check if the compression helps.

I needed to modify a test in test_obssurface.py because
a try-except block caused `to_netcdf` to be called twice,
and each call to `to_netcdf` asks for a uuid, but uuid4 was
mocked to only provide three values, which wasn't enough.
I added more fake uuids to fix this.