Created DatasourcePolicy class
DatasourcePolicy objects store:
- required_keys
- optional_keys
- min_required_keys

In addition, they have a method that takes
a metadata dictionary and returns the required metadata.

I created a dictionary with these objects for each
data type. This dictionary could easily be created from a JSON file.

I also added a function `get_required_metadata` that takes a data type
and a metadata dictionary, and returns the required metadata by looking
up the DatasourcePolicy object for that data type.