File hash tracking in new class, refactored search
File hashes are an "artefact" used by the pipeline for processing data.
They're not intrinsically part of the object store, so I've moved them
to their own class `FileHashTracker`.

This makes the pipeline code (currently in children of `BaseStore`)
slightly more complicated, but it makes the ObjectStoreConnection
code simpler.

I needed to refactor `search` to use the object store connection at
this point. I think this broke because `BaseStore`'s init function
was trying to load metadata. (I'm not exactly sure why this caused an
error, since `FileHashTracker` writes file hashes to the same location.)

TODO:
- further refactor of `search`
- change way that file hashes are stored (they need to be linked to a
UUID if we want to delete them)