Refactor to remove use of BaseStore context mgmt
The context manager for BaseStore children doesn't
do anything other than return an object now, so
I've removed (hopefully) all uses of the context
manager.

I also refactored tests to make it easier to
change the way that data is standardised by
putting the code for this in a fixture.

I created a `Standardiser` class in the standardise
module to help with the refactoring. This let's
us call `read_file` for a given bucket and data type
without directly referring to the BaseStore children.