Created `open_metastore` factory function
...which returns an instance of a child of BaseStore
given a data type and a bucket to connect to.

This will be used to refactor uses of BaseStore for accessing
metadata.

This function is different from `load_metastore` in the objectstore
module in that it only requires a bucket and data type, rather than
a bucket and key. So the interface it provides is more generic, and
means the user doesn't need to know about the internal organization of
the object store.