Removed 'bucket' parameter from standardise_*
We don't want the user API to mention 'bucket', so
the 'store' parameter is the only option for `standard_surface` etc
now.

I had to change the tests to reflect this change.

Currently, `standardise` still has 'bucket' as an argument
and the `standardise_*` functions use `get_writeable_bucket`
to get a bucket to pass to `standardise`