Made metastore a parameter in TinyDBMetaStore init
This is for two reasons:

1. we could modify how the TinyDB storage works so that changes are only
saved if all writes are successful. Currently, in `read_file`, we only
update the metastore after the `Datasource` is saved. A context manager
based on `load_metastore` opening a TinyDB that only writes when `close`
is called (and stages results when `write` is called) could do this in
a more natural way.

2. MongoDB supports sessions and transactions, and we might want to wrap calls to
a MongoDBMetaStore in a context manager that opens sessions/transactions.