Added DatasourceFactory class
This allows `ObjectStore` to create and load
`Datasource`s using only UUIDs. Any other info
needed to load a particular class of `Datasource`
is given to the `DatasourceFactory`.

This brings the treatment of the metastore and datasources
onto equal ground, since the particulars for initialising both
components must be established before passing them to an `ObjectStore`
instance.