Removed test case for `check_date`
I'm not sure why this test started failing with pandas 2.0 (i.e. why it
passed before).

`check_date` passes "1001" to `pd.Timestamp`, which converts it to
'01-01-1001 00:00:00', so the "check" passes.

The test expects "1001" to raise a `ValueError` (I think).

I just commented out the test case for now. The `check_date` function
is only used by aqmesh code, which I don't think is used anymore, so
perhaps we can just remove this function.