Added `force` option to `retrieve_atmospheric`
...and `ObsSurface.store_data`. If `force=True`, then
retrieved hashes will be ignored.

Two tests were added, one to check that if `force=False` (the default
option) then `retrieve_atmospheric` issues a warning that there is no
new data to process. If `force=True`, then this warning is not issued
(and the data is added again, although this is not explicitly checked).