Refactor: open_metastore now uses TinDBMetaStore
and `ClassicMetaStore` has been replaced by `DataClassMetaStore`,
since the data classes (children of `BaseStore`) were the only places
that `ClassicMetaStore` was being used without the `open_metastore`
function.

This means that we don't need a special class method for the data
classes, and it will allow us to add a `lock` method for data classes
to use in `assign_data`.