ModelScenario._get_data expands all lists
Previously there was code to expand a list of footprint inlet heights
into a list of dictionaries of search terms.

To add similar functionality for passing a list of object stores to
search for each data type, I made a function that expands dictionaries
with list values into a list of dictionaries with the same keys and
each combination of list items represented.

This adds the option to search over multiple stores and multiple fp
inlet heights. (As well as other lists of values...)