Allow ThreadSynchronizer to handle partial chunks
With new updates to xarray, it is no longer possible
to append data to a zarr store if the chunks of the
new data match the chunks of the existing data *and*
the existing data in the zarr store has a partial chunk
at the point where the new data will start.

To get around this, we can pass `safe_chunks=False`
to bypass the new xarray check and keep the existing
behavior.

In general, we should avoid this entire problem by
adding as much data as possible at once, using
`open_mfdataset` to open the new data files.

There are other ways around this issue besides
setting `safe_chunks=False`, but this would entail
detecting partial chunks in the region being written to,
and then rechuncking or splitting up the new data so that
the partial chunk is filled and then all the remaining
chunks of the new data align exactly with the chunks
in the zarr store.