Modified tests to work with pytest-xdist
Running `pytest -n 2` will run the tests in two parallel
processes. This roughly halves the time the tests take to run.

`pytest -n auto` will automatically choose the number of processess.
It seems that tests occasionally fail at random when there are many
processess, so it is probably best to run without the `-n` flag
before pushing a PR.