Added SimpleVersioning class
This class controls versions of a certain type of object.
To created a "versioned" class, the class you want to add versions to
must store its data in an attribute that can be overridden to point to
the current version.

Examples of this paradigm are given for versioned lists and versioned
dicts.

A simple "version type" class has been added to make it easier to use
versions of the form "v1", "v2", etc.