Combine metadata on retrieval, rather than storage
Metadata stored in the metastore and metadata stored by Datasources is
now combined when the object store is searched (or datasources are
retrieved from the object store).

The method of combination can be specified by passing a
`metadata_updater` function when creating an `ObjectStore` object.
By default, this does nothing, but when an object store is created via
the helper functions `open_object_store` and `locking_object_store`, we
pass a metadata updater that:
1. updates the datasource's metadata with any *new* metadata from the
metastore result
2. updates the metastore result with the datasource's metadata.

To do this correctly, we need to pass "extend keys" to the
`Datasource.add_metadata` method.

Another change required to avoid storing duplicate metadata on the
Datasource is that the "period" of the data is inferred in
`BaseStore.assign_data` and passed to `Datasource.add` as an argument.
The `get_period` method of `Datasource` was extracted to a helper
function in utils, and `Datasource.get_period` now calls this helper
function if it does not have a value for `.period` stored.

Similarly, we need to add the data type to the metadata that will be
stored in the metastore in `BaseStore.assign_data`, since previously,
this was added via the `Datasource`.

Previously, metadata stored in lists (like "tags") was combined using
`merge_and_extend_dict` in `Datasource.add_metadata`. The metastore
`.update` method has been extended to include this functionality now.