Use pint in ModelScenario calculations
The calculations for modelled obs and modelled baselines have been
updated to use pint when different types of data are combined.
This ensures that results like "footprint x flux" have consistent values
and units.

An option has been added to `footprints_data_merge` to set the output
units. If this is not specified, then the obs units are used.
Any variable starting with `mf` will be converted (if possible),
including `mf_mod`. Also `bc_mod` and `fp_x_flux` (and similar
variables) are converted to be in the specified output units.

When pint is used to preserve units, it is done by quantifying inside a
function and dequantifying before returning the result. This means that
users do not need to worry about using the `.pint` accessor for things
like `.pint.reindex` etc.

Besides the functions in `_modelled_obs.py` and `_modelled_baseline.py`,
the util function for stacking fluxes was changed to use pint. This is a
place where we probably could have just assigned the units `mol m-2 s-1`
to the result, but now we can handle fluxes with differing units, if
necessary. A set of tests was added for `stack_datasets`, which is the
helper function that required modification.