Updates for plotting, units
I updated the timeseries plotting to use pint to convert units. (With a
slight refactor to make it easier to do so.)

I removed the units conversion from `modelled_baselines`, since we can
convert with pint afterwards (or via `footprints_data_merge`).

I also made some changes so that the data types are not cast to float64,
which was happening with some pint quantified calculations.