Apply chunking schema to Flux
The `Flux` class now has a "chunking schema". This allows time resolved
flux data to be standardised, regardless of the length of the time
coordinate in the raw data.

An option to automatically scale the chunk sizes to fit within the
maximum (defaults to 300MB) was added, mostly because some flux tests
use the global edgar domain, which is very large.

It would make sense to chunk the lat/lon dimensions of the global edgar
domain (or any large domain), but that is somewhat beyond what I wanted
to achieve in this bug fix.