Removed warning about empty zarr store
If a store is empty, we treat it as a zarr store, for the purpose of
checking if a store is in zarr format, or OpenGHG <= 0.7 format.

This message runs every time you fetch data for an inversion, so it
really clutters up the inversion logs.

We could switch to `warnings.warn` and set a filter, but I've just opted
to remove the warning.