Fix `delete_version`
SimpleVersioning is used like a mixin class to modify
some *attribute* (like `.data` or `.store`) of another
class, say class `B`. When deleting a version, we want to
call the `delete` method of class `B`, not the `.delete` method of the
attribute (e.g. `.data` or `.store`), if it exists.