Remove "super init" option from SimpleVersioning
This feature means that the class that we want to add versioning to
usually needs to be modified to allow versioning to be added.

Removing this option means that the versioned class might need to
duplicate some code form the unversioned class's `__init__` method, but
this seems to be simpler than the alternative.