Fix bug by avoiding `Dataset.map`
Using `Dataset.map` applies the "remainder" function (usually
`mean_resample`) to any variables that haven't been resampled yet, but
it consideres each data variable separately, so behind the scenes,
`DataArrayResample` is used instead of `DatasetResample`, which means
that non-numeric variables are not dropped.

Applying the remainder function directly fixes this. I also made sure
that if there are no "remaining variables" that we don't try to resample
an empty Dataset.