Fixes from copilot review
Added `super().__init__` to `SimpleVersioning`. If kwargs are passed,
they are given to `super().__init__`; if kwargs is empty, any errors are
suppressed, so if `super().__init__` takes no arguments, it can still be
called.

This allows `VersionedMemoryStore` to use `super().__init__` to
initialise both `SimpleVersioning` and `MemoryStore`.

I didn't implement this for `VersionedZarrStore`, since this would
require modifying `ZarrStore` to be initialised without a zarr store
instance, which would require some unusual choices in `ZarrStore` (if we
take the perspective that the existence of `VersionedZarrStore`
shouldn't affect the superclass `ZarrStore`).